// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CatalogOption: {
            id: string;
            label: string;
            /** @enum {string} */
            status: "active" | "archived" | "unavailable";
        };
// ...
    };
// ...
}